Added radio buttons to filter dogs by gender
